requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-filter==23.1
django-cors-headers==4.0.0
python-decouple==3.8
Pillow==9.5.0
psycopg2-binary==2.9.6
celery==5.2.7
redis==4.5.5
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3
daphne==4.0.0
channels==4.0.0
channels-redis==4.1.0
